softap: show notification on client connect/disconnect
This commit builds and display a notification when a client/station
connect or disconnect on softap interface. This indication is send
via wificond module. On receveiving this, framework will query
dnsmasq to get device info and broadcast state change event via intents.

On softap teardown, this notification will be cleared.

Change-Id: I0f684049d93f5305f3ffb9647c2195f963efd64c