P2P: p2p flush on a group remove event.
On a group removal, the devise shall not be able to detect
the peer with which the group was formed earlier, as the
supplicant shall not give the P2P_DEV_FOUND indication for
an already reported one. Hence,do a p2p_flush to have a
better user experience in finding the p2p devices.
This will  get the device list entries in the framework in
sync with the suplicant peer device list.

CRs-Fixed: 962240
Change-Id: I13130f61120f7e9bd41824387364ca15535ea471