Wifi: Tear down STA/SAP separately for STA+SAP concurrent scenario
In case of STA + SAP Concurrency, cleanup need to be done separately
in Wifi Vendor Hal as well as wificond module. Also if one of STA
or SAP is active, avoid doing cleanup for other interface.

This commit adds separate APIs to teardownSta and teardownSap and
modifies Wifi State Machine(s) to do the cleanup separately.

CRs-Fixed: 2061689
Change-Id: Ia791e375fae504963bfb8c550e7f6936aa23a8cd