wifi: support FST in SoftAP mode
This change is a combination of the following commits:

1. wifi: unify fstman start/stop

fstman is now always started/stopped from wificond, both in softap and
supplicant mode. Use the same method (ctl.start/stop) in both cases.

2. WifiStateMachine: support FST in SoftAp mode

The handling of tethering was changed in recent android versions.
The Tethering class monitors WIFI_AP_STATE_CHANGED_ACTION, looks
at the interface name attached to the message and starts tethering
based on this interface.
The interface name is set in SoftApState.enter. When FST is enabled
it should be changed to bond0 so tethering will start on the correct
data interface for FST.
Note, there is a similar code in SoftApStateMachine which is used
when STA+SAP concurrency is enabled. Since we do not support FST
in this scenario, there is no need to modify the code in
SoftApStateMachine for now.

Change-Id: I16aeb54ddc237e5a77d14aefaa3ebe34f44e90bd