wifi: support FST in SoftAp mode with softap0 interface
Add support for running FST in SoftAp mode when the SoftAp
interface can be either wlan0 or softap0.
The change includes 2 parts:
1. During initialization when SoftAp interface is determined,
set it to property persist.vendor.fst.wifi.sap.interface.
2. When starting fstman, dynamically generate the fstman.ini
configuration file. The primary interface will be set to
wlan0 or softap0 based on the property set above.

Change-Id: Icb04976853420948194a7e2970a82bf66f660640