FILS: Reset AuthAlg to allow OPEN/FILS connection
When the target AP supports FILS(with ERP), reset the authAlg
configured to supplicant so that supplicant has the flexibility
to connect in OPEN mode when ERP keys are not available and in
FILS mode when ERP keys are available. Supplicant will have
freedom to choose when we don't configure any AuthAlg from
frameworks.

Change-Id: Ic84dff613442ff75aae3a8e648f4bd96e3e66d84
CRs-Fixed: 2099219