Wifi: Handle IP_REACHABILITY_LOST only for roaming scenarios
IP_REACHABILITY_LOST shall be triggered on a NUD_FAIL indication. The
current logic to check for the IP Reachability is to wait for the
response for the configured number of ARP requests to the gateway before
triggering a NUD_FAILURE. This feature when enabled by default is resulting
into some false positives (due to the ARP responses getting delayed from
the AP), leading to a disconnection / reconnection to the network.

Considering the pros / cons of this feature, this commit introduces a
new parameter to enable / disable this feature. If this feature is enabled,
the IP_REACHABILITY_LOST indication shall be honoured only for the first
10 seconds after the connect event. This shall ensure configured number of
ARP requests gets transmitted.

Change-Id: I64cbf358e8281556b5a02df26a62396996f88aca
CRs-Fixed: 2116615