Wifi: Create an interface for softap after the startHal.
It is this startHal , that configures the device in the respective
modes ( SoftAP / STA) . Also , this function loads the driver / configure
the firmware in the respective (STA / SAP ) modes. If the architectures
demand for creation of an interface for a SAP operation (in standalone
SAP, STA+SAP, SAP+SAP ) , the current behavior was creating the interface
before the startHal trigger. This will result in failures if driver is
not loaded

To address this, move creation of interface post startHal operation.

CRs-Fixed: 2121106
Change-Id: I0f19650752a474ccbd2e2ba0b171cf63ceb672a8