wifi: Remove dependency of init/cleanup ScannerImpl from ScanState Machines
ScannerImpl is initialized in the defaultState but the cleanup happens on the
exit of the StartedState. The cleanup of mScannerImpl also clears
mLastScanSettings which is the concern in the current context. Thus, for the
cases where the transition from the defaultstate to the StartedState doesn't
happen (maxBucket = 0 ), cleanup on the driver unload does not happen. This
leads to mLastScanSettings being not reset and hence the scan attempts after
the WiFi ON does not happen till it is cleared on a scan timeout. This leads
to a delay in the scan operations for a test scenario where the user
does a continuous WiFi OFF / ON operations looking for the scan performance.

Address this scenario, by not doing the initialization / cleanup of
mScannerImpl in a particular statemachine, but rather handle them in the
broadcast receiver context of WIFI_STATE_ENABLED/DISABLED events.

CRs-Fixed: 2127226
Change-Id: I2b6dc2978fb67ce56473e9dde9790b78dbbb180f