Wifi: Enhance debug logs in wifi framework
1. For dynamically adjusting the output of wifi framework log. When the user
turn on "Enable Wi-Fi Verbose Logging" in "Developer options", we will
check system property "vendor.qcom.wifi.debug", if the value is 1, the
related classes of wifi framework turn on its debug option. if the
value is 0, they are turned off. The related classes are
HalDeviceManager, WifiConnectivityManager, WifiController.
2. We just set the variable DBG to adjustable for WificondScannerImpl,
because it is controlled by WifiScanningService, it is not be affacted
by the option of "Enable Wi-Fi Verbose Logging". We can use the cammand
of "adb shell setprop log.tag.WificondScannerImpl VRBOSE" to change its
log level.
3. Fix uninitialized error for variable freqs.

CRs-Fixed: 2159526
Change-Id: Ia57c6ecb637db3e4b6ef5ef71ca15451d5e7bf3d