Wifi: Feature support for Repeater Mode
Repeater Mode is a feature where STA
network is shared to peers devices via
hotspot.
As part of this implementation if this
option is enabled then hotspot is
turned on using connected AP credentials.
Similarly when this option is tuned off,
then hotspot will turn on with already
preconfigured settings

CRs-Fixed: 2176779
Change-Id: Ifa999596210c548da7654a6fd012bbfd50908dc8