[WifiConfigManager] Log the disable reason when re-enabling.
A network may have been disabled a long time ago, and re-enabling it
wipes out the reason it was disabled in the first place. This can be
an important clue for diagnosing problems from a bug report, so log the
previous status and reason before re-enabling a network.

Bug: 149238341
Test: atest com.android.server.wifi
Test: manual - check that the logging is present dumpsys output:
      adb shell dumpsys wifi | grep -10 '^WifiConfigManager'

Change-Id: Id895918790a58e901b043bf607f83b30df9e8ebc
Merged-In: Id895918790a58e901b043bf607f83b30df9e8ebc